Start adding automated test suite in suite/.
The test suite is meant to be a start or step towards fully automated
and regression testing of this repository. This should make it easier
to make imrovements with confidence and help with identifying confusing
behavior or bugs in the current system.

The suite just scans all directories and runs all Run.hx in subdirectories.
There is one subdirectory of note called constructs. The goal with this
directory is to create an API and some consuming code and then to exercise
both codegen and the emitted externs to verify useful output. Currently,
there are two tests. A very basic class test and a failing enum test. I have
decided that the structure of this is enough to warrant publication, but
more tests need to be added.

Also, I am unsure of what testing frameworks are popular or desired.
Hopefully this is useful and can be adapted to any framework as desired.