suite: Add test demonstrating failure to export protected functions.
In haxe, private functions are like protected functions in OO. So
a consumer of an API should be able to override private functions
of a class exported by the API and access them through inheritence.

It would make sense to allow particular members to be marked as
not part of the public API and somehow hide them, but right now
codegen is not matching what haxe does.