Emit protected (haxe "private") fields in externs.
Fixes #2.

In haxe, there is no such thing as private in the traditional OO
sense. There is only protected. The only encapsulation supported
is preventing class instances of different types from accessing each
other's internals. Haxe always permits you to access internals
of super classes and to override methods (so this fixes overriding
non-public methods).

Another issue this fixes is the confusion that can result when Haxe
allows you to declare a function with the same name as a private
function in a super class. If you try to do that and the extern
includes the private function, Haxe will throw a compilation error
and require you to use override to indicate that you intended that.
If the super class *does* define such a function but it is omitted
from the extern, defining a function with the same name in the base
class effectively overrides it (at least on the JavaScript target).
I.e., these sorts of checks are only done by Haxe at compile time
and require the information to be available in the externs.